Refactor to make addEntry private
